places as edge from parent to child, all tests passing
